refactor: encapsulate getCloudinaryUrl and prefetch inside imageQueries.ts
- Moved `getCloudinaryUrl` out of a dedicated file directly into `imageQueries.ts`, allowing us to delete the redundant `cloudinary.ts` file entirely.
- Created `prefetchCloudinaryImage` inside `imageQueries.ts` to hide query client and url details from standard orchestrators.
- Updated `PieceDetail.tsx` to consume the new `prefetchCloudinaryImage` function, preventing it from having direct visibility or imports of `getCloudinaryUrl`.
- Updated mock structures in `ImageLightbox.test.tsx` and `PiecePhotoGalleryLightboxIntegration.test.tsx` to map `getCloudinaryUrl` correctly from `imageQueries.ts`.
- Removed `src/util/cloudinary.ts` registration in `web/BUILD.bazel`.
